scalaz-stream API for infinite streams
The underlying problem we're trying to solve is running effects on a
stream of events. We have a way of getting a finite Process of events
but we also need a way to convert that into an continual stream. This
provides a generic way to create infinite processes from finite streams.