Remove batchPut and replace with a sequence of puts
Using BatchWriteItem from DynamoDB caused us to lose the "absent"
constraint that eventsrc needs. Accidentally removing that constraint
meant that we could override historical events. This is really not good
in an event sourcing application.

We still have "batching" in the SaveAPI but we now do it by doing a
sequence of puts. We only do a single "latest" request and
optimistically try to write events from that. This gives us much fewer
reads when writing a lot of data.

LEM-1401